Move some accessiblity files in accessibility shared folder
