+++ src/shared/accessibility/getAccessibilityCategoryAndIcon.native.test.ts
