+++ src/shared/accessibility/getAccessibilityCategoryAndIcon.ts
