+++ src/shared/accessibility/getDetailedAccessibilityInfo.native.test.ts
