+++ src/ui/components/accessibility/AccessibilityAtom.native.test.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { render, screen } from 'tests/utils'
 
 import { AccessibilityAtom } from './AccessibilityAtom'
-import { HandicapCategory } from './getAccessibilityCategoryAndIcon'
+import { HandicapCategory } from '../../../shared/accessibility/getAccessibilityCategoryAndIcon'
 
 describe('AccessibilityAtom', () => {
   it('should display the validIcon when isValid is true', () => {
