+++ src/ui/components/accessibility/BasicAccessibilityInfo.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 import { OfferAccessibilityResponse } from 'api/gen'
 import { isNullOrUndefined } from 'shared/isNullOrUndefined/isNullOrUndefined'
 import { AccessibilityAtom } from 'ui/components/accessibility/AccessibilityAtom'
-import { HandicapCategory } from 'ui/components/accessibility/getAccessibilityCategoryAndIcon'
+import { HandicapCategory } from 'shared/accessibility/getAccessibilityCategoryAndIcon'
 import { Li } from 'ui/components/Li'
 import { Ul } from 'ui/components/Ul'
 import { getSpacing } from 'ui/theme'
